template data to use React Context
- Add TemplatesContext to manage template state centrally
- Update Sidebar and PageRecent to use useTemplates() hook
- Remove duplicate global handlers to prevent race conditions
- Aligns with existing React patterns
