+++ src/App.tsx
@@ -4,6 +4,7 @@ import { IntlProvider } from 'react-intl';
 import { getMessagesForLocale } from './localization/localization';
 import { LayoutContainer } from './components/LayoutContainer';
 import { SettingsProvider } from './components/SettingsContext';
+import { TemplatesProvider } from './components/TemplatesContext';
 
 const App = () => {
   const [locale, setLocale] = useState<Locale>("en");
@@ -25,7 +26,9 @@ const App = () => {
   return (
     <IntlProvider locale={locale} messages={messages}>
       <SettingsProvider>
-        <LayoutContainer id='homeContainer' />
+        <TemplatesProvider>
+          <LayoutContainer id='homeContainer' />
+        </TemplatesProvider>
       </SettingsProvider>
     </IntlProvider>
   );
